refactor(epub): 合并双份 EPUB 打包器到共享 EpubBuilder
TextToEpub 与 CuesToEpub 各自复制了整套 ZIP 打包器 + OCF/OPF/NCX/nav
生成器（~150 行逐字重复，仅 uid 前缀漂移）。抽取到
hibiki_audio/src/matching/epub_builder.dart 单一真相源，uid 前缀参数化
（hibiki- / hibiki-text- 保持不变，identity 判定不受影响）。

- 新增 TextToEpub 行为单测（此前该路径零覆盖）
- 测试侧 ZIP 读取器抽为共享 helpers/epub_zip_reader.dart
