+++ packages/hibiki_audio/lib/hibiki_audio.dart
@@ -34,3 +34,4 @@ export 'src/matching/collection_audio_matcher.dart';
 export 'src/matching/cue_file_index_assigner.dart';
 export 'src/matching/sasayaki_match_codec.dart';
 export 'src/matching/cues_to_epub.dart';
+export 'src/matching/epub_builder.dart';
