+++ packages/hibiki_audio/lib/src/matching/epub_builder.dart
@@ -1,0 +1,336 @@
+import 'dart:convert';
+import 'dart:io';
+import 'dart:typed_data';
+
+/// Shared EPUB 3 assembly infrastructure.
+///
+/// Single source of truth for the EPUB container pieces that are identical
+/// regardless of how chapter bodies are produced: the ZIP packer ([EpubZip]),
+/// the OCF/OPF/NCX/nav XML generators and XML escaping.
+///
+/// Used by [CuesToEpub] (cue spans) and the app-side `TextToEpub`
+/// (plain text / HTML / Markdown); both render their own chapter XHTML and
+/// hand the finished strings to [assemble].
+class EpubBuilder {
+  /// Packs [chapterXhtmls] plus all container metadata into EPUB bytes.
+  ///
+  /// [uidPrefix] is embedded verbatim before the timestamp in the
+  /// `dc:identifier` (e.g. `'hibiki-'`, `'hibiki-text-'`); callers rely on it
+  /// for identity/dedup of generated books, so it must stay stable.
+  static Uint8List assemble({
+    required String title,
+    required List<String> chapterXhtmls,
+    required String uidPrefix,
+    String? author,
+  }) {
+    final EpubZip zip = EpubZip();
+    final int chapterCount = chapterXhtmls.length;
+
+    // mimetype MUST be the first entry and stored (no compression) per EPUB spec.
+    zip.addStored('mimetype', utf8.encode('application/epub+zip'));
+
+    zip.addDeflated('META-INF/container.xml', utf8.encode(containerXml()));
+    zip.addDeflated(
+      'OEBPS/content.opf',
+      utf8.encode(contentOpf(
+        title: title,
+        author: author,
+        chapterCount: chapterCount,
+        uidPrefix: uidPrefix,
+      )),
+    );
+    zip.addDeflated(
+      'OEBPS/toc.ncx',
+      utf8.encode(tocNcx(title: title, chapterCount: chapterCount)),
+    );
+    zip.addDeflated(
+      'OEBPS/nav.xhtml',
+      utf8.encode(navXhtml(title: title, chapterCount: chapterCount)),
+    );
+
+    for (int i = 0; i < chapterCount; i++) {
+      zip.addDeflated(
+        'OEBPS/chapter-${i + 1}.xhtml',
+        utf8.encode(chapterXhtmls[i]),
+      );
+    }
+
+    return zip.build();
+  }
+
+  // ── XML / XHTML generators ────────────────────────────────────────────────
+
+  static String containerXml() => '<?xml version="1.0" encoding="UTF-8"?>\n'
+      '<container version="1.0"'
+      ' xmlns="urn:oasis:names:tc:opendocument:xmlns:container">\n'
+      '  <rootfiles>\n'
+      '    <rootfile full-path="OEBPS/content.opf"'
+      ' media-type="application/oebps-package+xml"/>\n'
+      '  </rootfiles>\n'
+      '</container>\n';
+
+  static String contentOpf({
+    required String title,
+    required int chapterCount,
+    required String uidPrefix,
+    String? author,
+  }) {
+    final String authorTag = (author != null && author.isNotEmpty)
+        ? '\n    <dc:creator>${escXml(author)}</dc:creator>'
+        : '';
+    final String now = DateTime.now()
+        .toUtc()
+        .toIso8601String()
+        .replaceFirst(RegExp(r'\.\d+Z$'), 'Z');
+
+    final StringBuffer manifest = StringBuffer();
+    final StringBuffer spine = StringBuffer();
+    for (int i = 1; i <= chapterCount; i++) {
+      manifest.write('    <item id="chapter-$i" href="chapter-$i.xhtml"'
+          ' media-type="application/xhtml+xml"/>\n');
+      spine.write('    <itemref idref="chapter-$i"/>\n');
+    }
+
+    final String uid = '$uidPrefix${DateTime.now().millisecondsSinceEpoch}';
+    return '<?xml version="1.0" encoding="UTF-8"?>\n'
+        '<package xmlns="http://www.idpf.org/2007/opf" version="3.0"\n'
+        '         unique-identifier="uid" xml:lang="ja">\n'
+        '  <metadata xmlns:dc="http://purl.org/dc/elements/1.1/">\n'
+        '    <dc:identifier id="uid">$uid</dc:identifier>\n'
+        '    <dc:title>${escXml(title)}</dc:title>$authorTag\n'
+        '    <dc:language>ja</dc:language>\n'
+        '    <meta property="dcterms:modified">$now</meta>\n'
+        '  </metadata>\n'
+        '  <manifest>\n'
+        '$manifest'
+        '    <item id="nav" href="nav.xhtml"'
+        ' media-type="application/xhtml+xml" properties="nav"/>\n'
+        '    <item id="ncx" href="toc.ncx"'
+        ' media-type="application/x-dtbncx+xml"/>\n'
+        '  </manifest>\n'
+        '  <spine toc="ncx">\n'
+        '$spine'
+        '  </spine>\n'
+        '</package>\n';
+  }
+
+  static String tocNcx({
+    required String title,
+    required int chapterCount,
+  }) {
+    final StringBuffer navPoints = StringBuffer();
+    for (int i = 1; i <= chapterCount; i++) {
+      navPoints.write(
+        '  <navPoint id="chapter-$i" playOrder="$i">\n'
+        '    <navLabel><text>Chapter $i</text></navLabel>\n'
+        '    <content src="chapter-$i.xhtml"/>\n'
+        '  </navPoint>\n',
+      );
+    }
+    return '<?xml version="1.0" encoding="UTF-8"?>\n'
+        '<!DOCTYPE ncx PUBLIC "-//NISO//DTD ncx 2005-1//EN"'
+        ' "http://www.daisy.org/z3986/2005/ncx-2005-1.dtd">\n'
+        '<ncx xmlns="http://www.daisy.org/z3986/2005/ncx/"'
+        ' version="2005-1">\n'
+        '  <head>\n'
+        '    <meta name="dtb:uid" content="hibiki-toc"/>\n'
+        '    <meta name="dtb:depth" content="1"/>\n'
+        '  </head>\n'
+        '  <docTitle><text>${escXml(title)}</text></docTitle>\n'
+        '  <navMap>\n'
+        '$navPoints'
+        '  </navMap>\n'
+        '</ncx>\n';
+  }
+
+  static String navXhtml({
+    required String title,
+    required int chapterCount,
+  }) {
+    final StringBuffer items = StringBuffer();
+    for (int i = 1; i <= chapterCount; i++) {
+      items.write('      <li><a href="chapter-$i.xhtml">Chapter $i</a></li>\n');
+    }
+    return '<?xml version="1.0" encoding="utf-8"?>\n'
+        '<!DOCTYPE html>\n'
+        '<html xmlns="http://www.w3.org/1999/xhtml"'
+        ' xmlns:epub="http://www.idpf.org/2007/ops" xml:lang="ja">\n'
+        '<head><meta charset="utf-8"/><title>${escXml(title)}</title></head>\n'
+        '<body>\n'
+        '  <nav epub:type="toc" id="toc">\n'
+        '    <ol>\n'
+        '$items'
+        '    </ol>\n'
+        '  </nav>\n'
+        '</body>\n'
+        '</html>\n';
+  }
+
+  // ── XML escaping ──────────────────────────────────────────────────────────
+
+  static String escXml(String s) => s
+      .replaceAll('&', '&amp;')
+      .replaceAll('<', '&lt;')
+      .replaceAll('>', '&gt;')
+      .replaceAll('"', '&quot;')
+      .replaceAll("'", '&apos;');
+}
+
+// ── Self-contained ZIP builder ───────────────────────────────────────────────
+//
+// Implements just enough of the ZIP spec (PKZIP 2.0, APPNOTE.TXT) for EPUB:
+//   • STORE  (method 0) for mimetype (required by EPUB spec)
+//   • DEFLATE (method 8) for all other files
+//
+// Uses dart:io ZLibCodec with raw=true for raw DEFLATE output.
+
+class EpubZip {
+  final List<_ZipEntry> _entries = [];
+
+  void addStored(String name, List<int> data) => _entries
+      .add(_ZipEntry(name: name, data: Uint8List.fromList(data), store: true));
+
+  void addDeflated(String name, List<int> data) => _entries
+      .add(_ZipEntry(name: name, data: Uint8List.fromList(data), store: false));
+
+  Uint8List build() {
+    final buf = BytesBuilder(copy: false);
+    final List<_LocalRecord> locals = [];
+
+    for (final entry in _entries) {
+      final int localOffset = buf.length;
+
+      final Uint8List nameBytes = Uint8List.fromList(utf8.encode(entry.name));
+      final int crc = _crc32(entry.data);
+
+      Uint8List compressed;
+      int method;
+      if (entry.store) {
+        compressed = entry.data;
+        method = 0; // STORE
+      } else {
+        // raw DEFLATE (wbits = -15)
+        compressed = Uint8List.fromList(
+          ZLibCodec(raw: true).encode(entry.data),
+        );
+        method = 8; // DEFLATE
+      }
+
+      // Local file header (signature 0x04034b50)
+      buf.add(_le32(0x04034b50));
+      buf.add(_le16(20)); // version needed: 2.0
+      buf.add(_le16(0)); // general purpose bit flag
+      buf.add(_le16(method));
+      buf.add(_le16(0)); // last mod time
+      buf.add(_le16(0)); // last mod date
+      buf.add(_le32(crc));
+      buf.add(_le32(compressed.length));
+      buf.add(_le32(entry.data.length));
+      buf.add(_le16(nameBytes.length));
+      buf.add(_le16(0)); // extra field length
+      buf.add(nameBytes);
+      buf.add(compressed);
+
+      locals.add(_LocalRecord(
+        nameBytes: nameBytes,
+        method: method,
+        crc: crc,
+        compressedSize: compressed.length,
+        uncompressedSize: entry.data.length,
+        localOffset: localOffset,
+      ));
+    }
+
+    // Central directory
+    final int cdOffset = buf.length;
+    for (final rec in locals) {
+      buf.add(_le32(0x02014b50)); // central directory signature
+      buf.add(_le16(20)); // version made by
+      buf.add(_le16(20)); // version needed
+      buf.add(_le16(0)); // general purpose bit flag
+      buf.add(_le16(rec.method));
+      buf.add(_le16(0)); // last mod time
+      buf.add(_le16(0)); // last mod date
+      buf.add(_le32(rec.crc));
+      buf.add(_le32(rec.compressedSize));
+      buf.add(_le32(rec.uncompressedSize));
+      buf.add(_le16(rec.nameBytes.length));
+      buf.add(_le16(0)); // extra field length
+      buf.add(_le16(0)); // file comment length
+      buf.add(_le16(0)); // disk number start
+      buf.add(_le16(0)); // internal file attributes
+      buf.add(_le32(0)); // external file attributes
+      buf.add(_le32(rec.localOffset));
+      buf.add(rec.nameBytes);
+    }
+    final int cdSize = buf.length - cdOffset;
+
+    // End of central directory record
+    buf.add(_le32(0x06054b50)); // signature
+    buf.add(_le16(0)); // disk number
+    buf.add(_le16(0)); // disk with start of central directory
+    buf.add(_le16(locals.length)); // entries on this disk
+    buf.add(_le16(locals.length)); // total entries
+    buf.add(_le32(cdSize));
+    buf.add(_le32(cdOffset));
+    buf.add(_le16(0)); // comment length
+
+    return buf.toBytes();
+  }
+
+  // ── CRC-32 (ISO 3309) ────────────────────────────────────────────────────
+
+  static final Uint32List _table = _buildCrcTable();
+
+  static Uint32List _buildCrcTable() {
+    final t = Uint32List(256);
+    for (int n = 0; n < 256; n++) {
+      int c = n;
+      for (int k = 0; k < 8; k++) {
+        c = (c & 1) != 0 ? 0xEDB88320 ^ (c >> 1) : c >> 1;
+      }
+      t[n] = c;
+    }
+    return t;
+  }
+
+  static int _crc32(Uint8List data) {
+    int crc = 0xFFFFFFFF;
+    for (final byte in data) {
+      crc = _table[(crc ^ byte) & 0xFF] ^ (crc >> 8);
+    }
+    return crc ^ 0xFFFFFFFF;
+  }
+
+  // ── Little-endian helpers ─────────────────────────────────────────────────
+
+  static Uint8List _le16(int v) =>
+      Uint8List(2)..buffer.asByteData().setUint16(0, v, Endian.little);
+
+  static Uint8List _le32(int v) =>
+      Uint8List(4)..buffer.asByteData().setUint32(0, v, Endian.little);
+}
+
+class _ZipEntry {
+  _ZipEntry({required this.name, required this.data, required this.store});
+  final String name;
+  final Uint8List data;
+  final bool store;
+}
+
+class _LocalRecord {
+  _LocalRecord({
+    required this.nameBytes,
+    required this.method,
+    required this.crc,
+    required this.compressedSize,
+    required this.uncompressedSize,
+    required this.localOffset,
+  });
+  final Uint8List nameBytes;
+  final int method;
+  final int crc;
+  final int compressedSize;
+  final int uncompressedSize;
+  final int localOffset;
+}
